add reth-trie-tests to prevent cyclic dep
