+++ crates/trie/trie/src/walker.rs
@@ -239,24 +239,17 @@ impl<C: TrieCursor> TrieWalker<C> {
 #[cfg(test)]
 mod tests {
     use crate::{
-        prefix_set::{PrefixSet, TriePrefixSets},
+        prefix_set::PrefixSet,
         trie_cursor::{
             noop::{NoopAccountTrieCursor, NoopStorageTrieCursor},
-            InMemoryTrieCursorFactory, TrieCursor,
+            TrieCursor,
         },
-        updates::TrieUpdates,
         walker::TrieWalker,
     };
-    use alloy_primitives::{hex, keccak256, Address, B256, U256};
-    use reth_db::{
-        test_utils::create_test_rw_db, transaction::DbTxMut, Database, DatabaseError,
-        HashedAccounts,
-    };
-    use reth_primitives::Account;
-    use reth_storage_errors::provider::ProviderError;
+    use alloy_primitives::hex;
+    use reth_storage_errors::db::DatabaseError;
     use reth_trie_common::{BranchNodeCompact, Nibbles, TrieMask};
-    use reth_trie_db::DatabaseTrieCursorFactory;
-    use std::{collections::HashMap, sync::Arc};
+    use std::collections::HashMap;
 
     struct MockCursor {
         // Internal state for the mock
@@ -371,7 +364,7 @@ mod tests {
         let changes = PrefixSet::default();
 
         let root_key = Nibbles::default();
-        let mut next_key = Nibbles::from_nibbles(&hex!("0d"));
+        let mut next_key = Nibbles::from_nibbles(hex!("0d"));
 
         // Root node with child at 0xd
         let root_node = BranchNodeCompact::new(
@@ -390,10 +383,9 @@ mod tests {
             None,
         );
 
-        mock_cursor
-            .expect_seek_exact(root_key.clone(), Some((root_key.clone(), root_node.clone())));
+        mock_cursor.expect_seek_exact(root_key.clone(), Some((root_key, root_node)));
 
-        mock_cursor.expect_seek(next_key.clone(), Some((next_key.clone(), level1_node.clone())));
+        mock_cursor.expect_seek(next_key.clone(), Some((next_key.clone(), level1_node)));
         mock_cursor.expect_current(Some(next_key.clone()));
 
         let mut walker = TrieWalker::new(mock_cursor, changes);
@@ -404,78 +396,4 @@ mod tests {
         next_key.push(0);
         assert_eq!(result, Some(next_key));
     }
-
-    #[test]
-    fn test_trie_walker_with_real_db_populated_account() {
-        let db = create_test_rw_db();
-
-        // Create test data
-        let address1 = Address::random();
-        let account1 = Account { nonce: 1, balance: U256::from(1000), bytecode_hash: None };
-
-        let tx = db.db().tx_mut().unwrap();
-
-        // Add account to the database
-        tx.put::<HashedAccounts>(B256::from(keccak256(address1)), account1)
-            .expect("Failed to insert account");
-
-        // Commit the transaction
-        tx.inner.commit().expect("Failed to commit transaction");
-
-        // Create a read transaction for testing
-        let tx = db.tx().expect("Failed to create transaction");
-
-        // Test account trie walker
-        let trie_updates = TrieUpdates::default();
-        let trie_nodes_sorted = Arc::new(trie_updates.into_sorted());
-
-        let trie_cursor_factory =
-            InMemoryTrieCursorFactory::new(DatabaseTrieCursorFactory::new(&tx), &trie_nodes_sorted);
-
-        let prefix_sets = TriePrefixSets::default();
-        let mut walker = TrieWalker::new(
-            trie_cursor_factory.account_trie_cursor().map_err(ProviderError::Database)?,
-            prefix_sets.account_prefix_set,
-        );
-
-        assert!(walker.key().is_some());
-        let result = walker.advance().expect("Failed to advance walker");
-        assert!(result.is_some());
-    }
-
-    /*
-        #[test]
-        fn test_trie_walker_with_real_db_populated_storage() {
-            let db = create_test_rw_db();
-
-            // Create test data
-            let address1 = Address::random();
-            let account1 = Account { nonce: 1, balance: U256::from(1000), bytecode_hash: None };
-
-            let tx = db.db().tx_mut().unwrap();
-
-            // Add storage entry
-            let storage_key = B256::random();
-            let storage_value = StorageEntry { key: storage_key, value: U256::from(42) };
-            tx.put::<HashedStorages>(B256::from(keccak256(address1)), storage_value)
-                .expect("Failed to insert storage");
-
-            // Commit the transaction
-            tx.inner.commit().expect("Failed to commit transaction");
-
-            // Create a read transaction for testing
-            let tx = db.tx().expect("Failed to create transaction");
-
-            let mut changes = PrefixSet::default();
-
-            // Test storage trie walker
-            let storage_cursor =
-                tx.cursor_read::<HashedStorages>().expect("Failed to create storage cursor");
-            let mut walker = TrieWalker::new(storage_cursor, changes);
-
-            assert!(walker.key().is_some());
-            let result = walker.advance().expect("Failed to advance walker");
-            assert!(result.is_some());
-    }
-        */
 }
